refactor: 스타일 파일 네이밍 통일 (style.ts -> styles.ts)
