+++ src/pages/OrderConfirmationPage/OrderConfirmationPage.tsx
@@ -2,20 +2,22 @@
 
 import { useNavigate } from 'react-router';
 import { useSelectedCartItemsContext } from '../../features/cart/context/useSelectedCartItemsContext';
-import { CartHeader, OrderPriceSummary } from '../../features/cart/ui';
+import { useOrderContext } from '../../features/order/context/useOrderContext';
+import { CartHeader } from '../../features/cart/ui';
 import CartItemInfo from '../../features/cart/ui/CartItemInfo';
-import OrderConfirmationText from '../../features/cart/ui/OrderConfirmationText';
+import { OrderConfirmationText, OrderPriceSummary } from '../../features/order/ui';
 import { ROUTES } from '../../shared/constants/routeConstants';
 import Navbar from '../../shared/ui/Navbar';
 import NavFooter from '../../shared/ui/NavFooter';
-import * as S from './OrderConfirmationPage.style';
+import * as S from './OrderConfirmationPage.styles';
 import Button from '../../shared/ui/Button';
 import CouponModal from '../../features/coupon/ui/CouponModal';
 import { useModal } from '@sanghee01/modal';
-import DeliveryInfo from '../../features/cart/ui/DeliveryInfo';
+import DeliveryInfo from '../../features/order/ui/DeliveryInfo';
 
 export default function OrderConfirmationPage() {
-  const { cartTypeQuantity, SelectedCartItems, updateRemoteArea } = useSelectedCartItemsContext();
+  const { SelectedCartItems } = useSelectedCartItemsContext();
+  const { cartTypeQuantity, updateRemoteArea } = useOrderContext();
   const {
     isOpen: isCouponModalOpen,
     handleOpen: handleOpenCouponModal,
